date issue fix in fxvanilla options in prices and code refactoring in optionchains
